% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma.spec2.R
\name{arma.spec2}
\alias{arma.spec2}
\title{Modified astsa::arma.spec}
\usage{
arma.spec2(
  ar = 0,
  ma = 0,
  var.noise = 1,
  n.freq = 500,
  plot = FALSE,
  main = "from specified model",
  frequency = 1,
  ylim = NULL,
  ...
)
}
\arguments{
\item{ar}{vector of AR parameters}

\item{ma}{vector of MA parameters}

\item{var.noise}{variance of the noise}

\item{n.freq}{number of frequencies}

\item{frequency}{for seasonal models, adjusts the frequency scale}

\item{...}{additional graphical parameters}
}
\value{
frequencies and spectral ordinates of inputted ARMA process
}
\description{
Same as astsa::arma.spec, but plot is optional
}
