% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgram.R
\name{pgram}
\alias{pgram}
\title{Returns the periodogram of a signal}
\usage{
pgram(x, plot = TRUE, ...)
}
\arguments{
\item{x}{Data}

\item{plot}{If TRUE (default), a graph is produced. If FALSE, no graph is 
produced.}

\item{...}{Additional arguments passed to \code{plot}}
}
\description{
computes and plots the periodogram of a signal. The periodogram
is an estimate of the spectral density of a signal and is used to examine
the dominant frequencies present in the data.
}
\examples{
pgram(rnorm(100))
pgram(sin(2*pi*0.1*0:100) + 3*cos(2*pi*0.33*0:100))

}
\author{
Jared Fisher
}
