% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima.compare.R
\name{sarima.compare}
\alias{sarima.compare}
\title{Compare the behavior of multiple arima models to a reference model's residuals}
\usage{
sarima.compare(
  x,
  models = c(p = 0, d = 0, q = 1),
  ref = c(p = 0, d = 0, q = 0),
  max.lag = 15,
  plot = TRUE,
  model.colors = c("dodgerblue", "hotpink")
)
}
\arguments{
\item{x}{data}

\item{ref}{Order of reference ARIMA model whose residuals will be used for comparison}

\item{max.lag}{number of lags to compute in the ACF/PACF and 
Ljung-Box Statistic plots.}

\item{plot}{plots diagnostic}

\item{model}{List of vectors of ARIMA(p,d,q) model orders.}
}
\value{
For the fitted ARIMA models, list of theoretical ACF/PACF values, 
theoretical spectral density values, and Ljung-Box Statistic p-values.
}
\description{
Fits ARIMA(p,d,q) models to the data and plots their ACF/PACF, 
Ljung-Box Statistic p-values, and spectral density. Also plots for comparison
the residuals of a reference ARIMA model.
}
