% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima.compare1.R
\name{sarima.compare1}
\alias{sarima.compare1}
\title{Diagnostic plots for a SARIMA model}
\usage{
sarima.compare1(
  x,
  model = c(p = 0, d = 0, q = 1),
  ref = c(p = 0, d = 0, q = 0),
  max.lag = 15,
  plot = TRUE
)
}
\arguments{
\item{x}{data}

\item{model}{Vectors of ARIMA(p,d,q) or SARIMA(p,d,q, P,D,Q,S) model order.}

\item{ref}{reference model.}

\item{max.lag}{number of lags to compute in the ACF/PACF and 
Ljung-Box Statistic plots.}

\item{plot}{plots diagnostic}
}
\value{
list of fitted ACF, PACF, Ljung-Box Statistic p-values, and spectral
density of the fitted ARIMA model.
}
\description{
Plots ACF/PACF, Ljung-Box Statistic p-values, and spectral density of the
ARIMA(p,d,q) or SARIMA(p,d,q, P,D,Q,S) process fitted to the inputted data.
}
