% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima.cv.R
\name{sarima.cv}
\alias{sarima.cv}
\title{time series cross validation on multiple SARIMA models}
\usage{
sarima.cv(x, models = c(p = 0, d = 0, q = 1), n.frames = 10, n.points = 5)
}
\arguments{
\item{x}{data}
}
\description{
time series cross validation on multiple SARIMA models
}
